ŝanĝis la manieron laŭ kiu la fenestroj estas bildigitaj (kiu solvis la problemon #1), oni nun povas selekti la programon el la kunteksta menuo de selektita teksto

+++ firefox/public/inject.js
@@ -3,7 +3,7 @@
 (() => {
     let relativeDiv;
     const body = document.body;
-    const head = document.head;
+    let root;
     let turnedOnLocally = true;
     let backgroundLocal = { color: "#f8ebe8" };
     let textLocal = { color: "#000", size: 15 };
@@ -11,59 +11,67 @@
     const createStyle = () => {
         const style = document.createElement("style");
         style.textContent = `
+    html, body, div, span, applet, object, iframe,
+    h1, h2, h3, h4, h5, h6, p, blockquote, pre,
+    a, abbr, acronym, address, big, cite, code,
+    del, dfn, em, img, ins, kbd, q, s, samp,
+    small, strike, strong, sub, sup, tt, var,
+    b, u, i, center,
+    dl, dt, dd, ol, ul, li,
+    fieldset, form, label, legend,
+    table, caption, tbody, tfoot, thead, tr, th, td,
+    article, aside, canvas, details, embed, 
+    figure, figcaption, footer, header, hgroup, 
+    menu, nav, output, ruby, section, summary,
+    time, mark, audio, video {
+      margin: 0;
+      padding: 0;
+      border: 0;
+      font-size: 100%;
+      font: inherit;
+      vertical-align: baseline;
+    }
+    /* HTML5 display-role reset for older browsers */
+    article, aside, details, figcaption, figure, 
+    footer, header, hgroup, menu, nav, section {
+      display: block;
+    }
+    body {
+      line-height: 1;
+    }
+    ol, ul {
+      list-style: none;
+    }
+    blockquote, q {
+      quotes: none;
+    }
+    blockquote:before, blockquote:after,
+    q:before, q:after {
+      content: '';
+      content: none;
+    }
+    table {
+      border-collapse: collapse;
+      border-spacing: 0;
+    }
+    .PIValdonajxo * {
+      all: revert !important;
+    }
     .PIValdonajxo {
       z-index: 2147483647 !important;
       box-sizing: border-box !important;
-      font-family: "Open Sans", sans-serif !important;
-      line-height: 1.25em !important;
-      vertical-align: initial !important;
       position: absolute !important;
       overflow-y: auto !important;
       border-style: solid !important;
       width: 300px !important;
       height: 300px !important;
       padding: 10px !important;
-    }
-    .PIValdonajxo * {
-      font: revert !important;
-    }
-    .PIValdonajxo-w-definitions *:not(sub, sup) {
-      opacity: 1 !important;
-      display: inline !important;
-      visibility: visible !important;
-      position: relative !important;
-      line-height: 1.25em !important;
-      vertical-align: initial !important;
-    }
-    .PIValdonajxo-w-definitions sup, .PIValdonajxo sub {
-      opacity: 1 !important;
-      display: inline !important;
-      visibility: visible !important;
-      position: relative !important;
-      font-size: smaller !important;
-      line-height: 1.25em !important;
-      vertical-align: baseline !important;
-    }
-    .PIValdonajxo-w-definitions sup {
-      top: -0.5em !important;
-    }  
-    .PIValdonajxo-w-definitions sub {
-      bottom: -0.5em !important;
-    }
-    .PIValdonajxo__link {
-      text-decoration: underline !important;
-      cursor: pointer !important;
-    }
-    .PIValdonajxo__strong {
-      display: block !important;
-    }
-    .PIValdonajxo__list {
-      list-style: initial !important;
+      font-family: 'Open-sans', sans-serif !important;
+      line-height: 1.5 !important;
     }
   `;
-        head.appendChild(style);
+        root.appendChild(style);
     };
-    createStyle();
     const createRelativeDiv = () => {
         relativeDiv = document.createElement("div");
         Object.assign(relativeDiv.style, {
@@ -71,7 +79,11 @@
             top: "0",
             left: "0",
         });
+        relativeDiv.classList.add("PIValdonajxo-root");
         body.appendChild(relativeDiv);
+        relativeDiv.attachShadow({ mode: "open" });
+        root = relativeDiv.shadowRoot;
+        createStyle();
     };
     createRelativeDiv();
     const addDerivationsLinks = (div, results) => {
@@ -140,7 +152,7 @@
         const URLaddress = "https://vortaro.net/py/serchi.py?m=" + query;
         return browser.runtime
             .sendMessage({ URLaddress })
-            .then((response) => (Object.assign(Object.assign({}, response), { word })));
+            .then((response) => ({ ...response, word }));
     };
     const addWaitingMessage = (div) => {
         const strong = document.createElement("strong");
@@ -184,6 +196,9 @@
         ? rangeRect.left - divRect.left - 300 + rangeRect.width
         : rangeRect.left - divRect.left;
     const createDiv = (range, number, word) => {
+        if (!document.querySelector(".PIValdonajxo-root")) {
+            createRelativeDiv();
+        }
         const divRect = relativeDiv.getBoundingClientRect();
         const rangeRect = range.getBoundingClientRect();
         const div = document.createElement("div");
@@ -201,27 +216,20 @@
         div.className = "PIValdonajxo";
         div.setAttribute("number", String(number));
         addWaitingMessage(div);
-        body.appendChild(div);
+        root.appendChild(div);
         const responsePromise = makeRequest(word);
         responsePromise.then((response) => response ? addContentToDiv(response, div) : addConnectionMessage(div));
-    };
-    const generateNumber = (target) => {
-        var _a;
-        const addonElements = [
-            ...document.querySelectorAll(".PIValdonajxo"),
-        ];
-        return (Number((_a = addonElements
-            .find((element) => element.contains(target))) === null || _a === void 0 ? void 0 : _a.getAttribute("number")) + 1 || 1);
     };
     document.addEventListener("mouseup", (e) => {
         const selection = window.getSelection();
         const element = e.target;
+        const originalTarget = e.composedPath()[0];
+        const number = element === root.host ? getNumber(originalTarget) + 1 : 1;
         if (element.classList.contains("PIValdonajxo__link")) {
             selection.removeAllRanges();
             let range = document.createRange();
             range.selectNode(element);
             selection.addRange(range);
-            const number = generateNumber(element);
             createDiv(range, number, element.innerText);
             selection.removeRange(range);
         }
@@ -234,8 +242,7 @@
                 turnedOnLocally &&
                 /^[À-ž\w\d\s.,\-;]+$/.test(word) &&
                 !isInCont) {
-                const number = generateNumber(element);
-                if (!document.querySelector('.PIValdonajxo[number="' + number + '"]')) {
+                if (!root.querySelector('.PIValdonajxo[number="' + number + '"]')) {
                     createDiv(selection.getRangeAt(0), number, word);
                 }
             }
@@ -256,25 +263,48 @@
             selection.removeAllRanges();
         }
     };
-    const getNumber = (elements, target) => {
-        var _a;
-        return Number((_a = elements
-            .find((element) => element.contains(target))) === null || _a === void 0 ? void 0 : _a.getAttribute("number")) || 0;
+    const getNumber = (target) => {
+        let element = target;
+        while (!element.classList.contains("PIValdonajxo")) {
+            element = element.parentElement;
+        }
+        return +element.getAttribute("number");
     };
     document.addEventListener("mousedown", (e) => {
         const contElements = [...document.querySelectorAll("[contenteditable]")];
+        const originalTarget = e.composedPath()[0];
         const element = e.target;
         const isInCont = contElements.some((contElement) => contElement.contains(element));
         if (e.button === 0 && turnedOnLocally && !isInCont) {
-            const addonElements = [...document.querySelectorAll(".PIValdonajxo")];
-            const bound = getNumber(addonElements, element);
-            if (bound > 0) {
+            const addonElements = [...root.querySelectorAll(".PIValdonajxo")];
+            if (element === root.host) {
+                const bound = getNumber(originalTarget);
                 deleteSubelements(addonElements, bound);
             }
             else {
                 deleteAllElements(addonElements);
             }
             removeSelection();
+        }
+    });
+    browser.runtime.onMessage.addListener((message, sender, sendResponse) => {
+        let selection = window.getSelection();
+        let focusNode = selection.focusNode;
+        let number = 1;
+        const addonElements = [...root.querySelectorAll(".PIValdonajxo")];
+        const word = selection.toString().trim().toLowerCase();
+        const contElements = [...document.querySelectorAll("[contenteditable]")];
+        const isInCont = contElements.some((contElement) => contElement.contains(focusNode.parentElement));
+        if (turnedOnLocally && /^[À-ž\w\d\s.,\-;]+$/.test(word) && !isInCont) {
+            if (root.contains(focusNode.parentElement)) {
+                const bound = getNumber(focusNode.parentElement);
+                number = bound + 1;
+                deleteSubelements(addonElements, bound);
+            }
+            else {
+                deleteAllElements(addonElements);
+            }
+            createDiv(selection.getRangeAt(0), number, word);
         }
     });
     browser.storage.sync
@@ -289,24 +319,26 @@
             borderLocal = border;
         }
     });
-    browser.storage.onChanged.addListener(({ turnedOn }) => {
-        turnedOnLocally = turnedOn.newValue;
-        if (turnedOnLocally === false) {
-            const addonElements = document.querySelectorAll(".PIValdonajxo");
-            deleteAllElements(addonElements);
+    browser.storage.onChanged.addListener(({ turnedOn, background, text, border }) => {
+        if (turnedOn) {
+            turnedOnLocally = turnedOn.newValue;
+            if (turnedOnLocally === false) {
+                const addonElements = root.querySelectorAll(".PIValdonajxo");
+                deleteAllElements(addonElements);
+            }
         }
-    });
-    browser.storage.onChanged.addListener(({ background, text, border }) => {
-        backgroundLocal = background.newValue;
-        textLocal = text.newValue;
-        borderLocal = border.newValue;
-        const addonElements = document.querySelectorAll(".PIValdonajxo");
+        if (background)
+            backgroundLocal = background.newValue;
+        if (text)
+            textLocal = text.newValue;
+        if (border)
+            borderLocal = border.newValue;
+        const addonElements = root.querySelectorAll(".PIValdonajxo");
         addonElements.forEach((element) => {
             Object.assign(element.style, {
                 background: backgroundLocal.color,
                 "border-width": borderLocal.width + "px",
                 "border-color": borderLocal.color,
-                "border-style": "solid",
                 fontSize: textLocal.size + "px",
                 color: textLocal.color,
             });
